moved optiontypes to a new model file
